refactor: Create a dedicated directory and package for API endpoints handlers
